+++ frb_codegen/src/library/codegen/generator/wire/rust/spec_generator/misc/ty/dart_fn.rs
@@ -2,7 +2,7 @@ use crate::codegen::generator::acc::Acc;
 use crate::codegen::generator::wire::rust::spec_generator::base::*;
 use crate::codegen::generator::wire::rust::spec_generator::misc::ty::WireRustGeneratorMiscTrait;
 use crate::codegen::generator::wire::rust::spec_generator::output_code::WireRustOutputCode;
-use crate::codegen::ir::result::IrDartFnOutput;
+use crate::codegen::ir::ty::dart_fn::IrDartFnOutput;
 use crate::codegen::ir::ty::IrTypeTrait;
 use crate::library::misc::consts::HANDLER_NAME;
 use itertools::Itertools;
