+++ frb_codegen/src/library/codegen/ir/ty/dart_fn.rs
@@ -1,4 +1,3 @@
-use crate::codegen::ir::result::IrDartFnOutput;
 use crate::codegen::ir::ty::dart_opaque::IrTypeDartOpaque;
 use crate::codegen::ir::ty::{IrContext, IrType, IrTypeTrait};
 use itertools::Itertools;
@@ -7,6 +6,11 @@ crate::ir! {
 pub struct IrTypeDartFn {
     pub inputs: Vec<IrType>,
     pub output: Box<IrDartFnOutput>,
+}
+
+pub(crate) struct IrDartFnOutput {
+    pub(crate) normal: IrType,
+    pub(crate) error: Option<IrType>,
 }
 }
 
@@ -45,5 +49,40 @@ impl IrTypeTrait for IrTypeDartFn {
 impl IrTypeDartFn {
     pub(crate) fn get_delegate(&self) -> IrType {
         IrType::DartOpaque(IrTypeDartOpaque)
+    }
+}
+
+impl IrDartFnOutput {
+    pub(crate) fn visit_types<F: FnMut(&IrType) -> bool>(
+        &self,
+        f: &mut F,
+        ir_context: &impl IrContext,
+    ) {
+        self.normal.visit_types(f, ir_context);
+        if let Some(error) = &self.error {
+            error.visit_types(f, ir_context);
+        }
+    }
+
+    pub(crate) fn safe_ident(&self) -> String {
+        format!(
+            "{}_{}",
+            self.normal.safe_ident(),
+            self.error
+                .map(|x| x.safe_ident())
+                .unwrap_or("None".to_owned())
+        )
+    }
+
+    pub(crate) fn rust_api_type(&self) -> String {
+        if let Some(error) = &self.error {
+            format!(
+                "std::result::Result<{}, {}>",
+                self.normal.rust_api_type(),
+                error.rust_api_type()
+            )
+        } else {
+            self.normal.rust_api_type()
+        }
     }
 }
