+++ frb_codegen/src/library/codegen/parser/type_parser/dart_fn.rs
@@ -1,7 +1,7 @@
 use crate::codegen::ir::field::IrField;
 use crate::codegen::ir::ident::IrIdent;
 use crate::codegen::ir::namespace::{Namespace, NamespacedName};
-use crate::codegen::ir::result::IrDartFnOutput;
+use crate::codegen::ir::ty::dart_fn::IrDartFnOutput;
 use crate::codegen::ir::ty::dart_fn::IrTypeDartFn;
 use crate::codegen::ir::ty::enumeration::{
     IrEnum, IrEnumIdent, IrEnumMode, IrTypeEnumRef, IrVariant, IrVariantKind,
